Fix head pointer misalignment on re-use.
When reusing blocks of memory, the memory manager would incorrectly hand out a pointer to the first free byte of the cache, rather than the first byte of the reusable block. In these cases, it also failed to increment the cache pointer, leading to the same blocks of memory being handed out to multiple recipients.