Deprecate the OpenGL frame surface
I can't properly test the OpenGL surface in my virtual machines, and
it's not clear that this feature is even needed. It's best not to let it
litter the codebase, and remove it.

The OpenGL mode was added after a user reported low frame rates on their
low-end CPU. There was never a follow-up on whether the mode helped this
person; but from what I've seen in myvirtual machines, it's possible that
it wasn't a considerable help.