Force x264vfw to output via the VFW interface
The x264vfw codec gives the user two options for output: the VFW
interface, or directly into file. VCS can only work with x264vfw if it
outputs via the VFW interface; so force it to always do so when VCS is
recording, by forcing the option in x264vfw's registry settings.