GUI: Add option to set aspect ratio mode
The aspect ratio mode lets the user specify how frames should be padded
to maintain the desired aspect ratio. Some non-4:3 DOS resolutions, for
instance, used to be compressed to 4:3 for display on monitors of the
time, and a user may still want to replicate that behavior. The aspect
mode gives them a way to do so.