Reorganize capture.cpp
Functions to do with the capture hardware were moved into a struct
dedicated to such functions, so that they're not scattered all over the
place.

Many static parameters in capture.cpp were removed in favor of polling
them from the capture API every time access to them is desired. This
means less getters in capture.cpp, and less fiddly state management. The
extent to which performance is affected is to some extent dependent on
the API, but based on brief initial testing, it shouldn't be dramatic,
at least.