Continue globally reorganizing capture.cpp
Functions related to managing the state of the capture - like starting
and stopping - were moved together into a struct.

For consistency, the RGBEasy API callback functions were grouped into a
struct instead of the namespace they were in previously.

Functions dealing with capture both in and outside of capture.cpp that
contained *_input_* in their name - referring to capturing - were
renamed to *_capture_*.