GUI: Make interface style selector a drop-down box
So in theory, more than one custom style can be made available, which
wasn't the case when the selector was just a checkbox.

The whole custom styling thing still needs work. Right now, it's all
hard-coded, but ideally I want the user to have the ability to provide
their own custom styles, too, without having to recompile the program.