Fix F5 hotkey not updating custom video params
The effect of this bug was that you could adjust the frame's position
using the F5 key, save these new parameters, and the change wouldn't be
registered in the save file, since the custom parameter objects - which
are the things that actually get saved - didn't get updated to reflect
the changes produced by the F5 key.

The current system for dealing with custom video parameters isn't too
good, and I want to work on it a bit in the future. This type of bug
shouldn't be allowed to occur.