Fix scaler not ignoring disabled aspect correction
When the input and output frames are the same size, and aspect correction
is either disabled or 'native', no scaling needs to be done. However, the
code was failing to test whether aspect correction was disabled, and was
instead just testing for whether the mode was 'native'.

The bug resulted in some wasted compute, but probably had no effect
otherwise.