Add #includes missing on newer build chains
It was reported that building fails on Fedora 30. This commit adds
includes that were missing there; and hopefully solves any similar issues
that may exist when building VCS on other such newer distros.

Note that the OpenCV and RGBEASY code paths were not tested -
they may or may not still contain unsolved dependencies.