Filter graph: Implement loading of legacy filter sets
When saved filter sets of the previous version of VCS are loaded in,
they're converted internally into the new filter graph format.

There are a three things to be aware. First, the legacy filter system
allowed the user to specify whether to apply a filter before the
frame was scaled or afterwards; whereas the node graph only allows
pre-scaling filtering. The conversion thus flattens this distinction.
Second, the legacy filtering system allowed the user to specify which
frame scaler to use with that set of filters; whereas the filter graph
does not allow this option. This scaler selection will be discarded upon
conversion. Third, the legacy file-loading code still contains some
obsolete code paths that cause unwanted side effects - this commit does
not yet bring the repo to a release-ready state.

The original save file will not be modified - the conversion operates
solely on data copied into memory.