Add program-wide dev version toggle
While it's useful to have much the same experience using the dev
version as the release version, it's nonetheless handy to be able to
enable/disable certain features depending on whether this is a dev
build or a release build.

This commit adds the program-wide DEV_VERSION boolean, which can be used
to set the program into developer mode. (But deviation of the dev build
from the release build should be kept to a minimum.)