Redesign the video parameters dialog
Instead of going with a graph approach that I sketched out in a previous
commit, I've kept the normal GUI and introduced presets - each preset
contains a set of custom video parameters as well as one or more user-
definable conditions for when the preset's parameters should apply.

Note that this is still work in progress and certain parts of the
dialog aren't yet implemented.