Replace kpropagate_xxxx with an event system
Instead of the hard-coded event propagation of kpropagate_xxxx,
the new event system provides subscription-based event propagation.

A given subsystem of VCS wishing to be notified of an event will
register a callback function with that event, the function being
called by the event whenever the event occurs.

This is work in progress and fairly untested as of now.