memory: Increase memory cache size to 1 GB
With the recent introduction of a frame queue to the video recorder,
the memory cache needs to be a bit bigger.

This is the same size as used by the most recent release version of
VCS; i.e. it was lowered in the current developmental version and
is now restored to its original value.