capture/video4linux: Re-initialize input channel every time...
...the source video mode changes.

It seemed difficult to get an active input channel to dynamically
resize its capture buffers to reflect changes to the source video mode.
So let's just re-initialize the whole input channel on mode changes,
that appears to work.

With this change, the capture buffers will be only as large in
resolution as needed to capture the current source signal; whereas
before, the capture buffers were set on initialization to be large
enough to store the largest supported capture resolution. The change
improves capture performance considerably.

Since this commit alters core capture logic, regressions can be
expected.