capture/video4linux: Don't try to avoid unnecessary ioctl()...
...calls when setting video parameters.

Previously, the system wouldn't re-assign a parameter if it thought
the parameter already had that value. But there's a bug preventing the
cache of known values from updating under certain circumstances, and I
don't feel like debugging it. It's probably fine to make a few more
calls to ioctl() anyway.