anti-tear: Optimize tear scanner (one tear case)
Instead of scanning the entire vertical extent of the frame for the
start of the tear, we'll recursively split the range into half, then
half again, etc., until we find the row on which the tear starts. This
way, we only need to test a fraction of the frame's rows.

This optimization is only for the "up to one tear per frame" case.