anti-tear: Add custom scan direction
Most common video modes seem to redraw the screen from bottom to top,
but some do it from top to bottom. The anti-tearer's scan direction
needs to be set accordingly: down for sources that redraw from bottom
to top, and up for the other ones.

The prior implementation - without a customizable direction - always
scanned down.