filter: Add anti-tear as a filter
Anti-tearing can now be used in one of two ways: either via the anti-
tear dialog (Ctrl+A), or with this new filter via the filter graph
dialog (Ctrl+F). The dialog way and the filter way are independent of
each other.

The benefit of using the filter is that it lets you save multiple anti-
tearing configurations (by saving the filter graph), whereas the dialog
gives you only one configuration. On the other hand, the dialog may be
quicker to use, since you don't need to wire up a filter.