capture/vision_v4l: Improve handling of...
...resolution switching and the 'no signal' condition.

The capture API should now correctly detect when the capture signal is
lost; although there's some uncertainty about whether this
implementation is portable (the V4L "signal_type" control's id has been
hard-coded as 0x8000013, but I'm not sure whether it corresponds to that
id only on my system or with the Vision driver in general).

This commit does introduce some hacky, kludgy code, but hey.