docs/user-manual: Update manual to use dokki's intermediate format
In the intermediate format, each section of the manual is defined in its
own Markdown file, and the files are built into a final HTML
distributable (index.html) using dokki's intermediate compiler.

The intermediate format is work in progress in the dokki project;
instructions for building the final distributable of this manual will be
added later.