docs/user-manual: Add ability to select manual's version
The base version of the manual is now targeted at the VCS repo's
master branch, including documentation for features that aren't yet
available in a binary release.

Outside of the repo, I host snapshots of the manual for the various
point releases of VCS, which this new version selector widget lets the
user switch between.