gui/filter-graph: Tweak node styling some more
I've only tested the changes of this and the previous commit in Linux,
not in Windows, though have seen this kind of Qt GraphicsScene stuff wig
out more on Windows than Linux, so it's possible the changes broke
some visual continuity on the Windows side. I'll have a look later.