gui/filter-graph: Select (only) clicked node on right click
If multiple nodes are selected when one of them is right-clicked, the
others are removed from the selection.

Ideally there would be a separate right-click menu for when the
selection includes multiple nodes - so you can e.g. set the background
color on multiple nodes at a time - but for now we don't have that.