anti-tear: Demote an assertion to a debug message
Something is causing copy_frame_pixel_rows() to get incorrect values for
fromRow/toRow, triggering the assertion. It seems to happen when the
video mode changes, but only sometimes.

I don't feel like debugging it right now, so let's just demote the
program-ending assertion to a debug message, and get to it later.