Add a suppression file requirement for AddressSanitizer
On my Ubuntu 20.04 there's a few third-party libraries with memory leaks
that trip AddressSanitizer, so it needs a suppression file.

What you need to put into the file - if anything - depends on your system,
but for me this does the trick:

leak:/usr/lib/x86_64-linux-gnu/libfontconfig.so
leak:/usr/lib/x86_64-linux-gnu/libglib-2.0.so

Annoyingly, the suppression file needs to be located relative to the VCS
executable, so there doesn't seem to be a portable way to pre-include it
in the repo, and so you need to manually create it even if you don't
intend to use suppressions.

The suppression file isn't needed for release builds, as they don't use
AddressSanitizer.