Reimplement capture device interface
The capture device interface is now more generic, using key-value pairs
instead of functions to communicate between VCS and the device. This
allows VCS to be decoupled more strongly from Datapath's capture devices
and so (in theory) make it easier to add other kinds of capture sources.

This reimplementation liberally culls the number of superfluous capture
features. For example, the UI no longer provides metadata about the
capture device. Sone of these features might be put back in later.

The new interface hasn't been tested on an actual capture card yet and
so is expected to require some fixing in future commits.