ui/video-presets: Fix locked preset not activating reliably
Fixes a few bugs associated with locked presets. Restoring the selected
preset index is no longer dependent on the order of the presets in the
GUI, the output window's title bar is now updated with the title of the
locked preset when VCS starts up, and locked presets now apply correctly
when VCS is started while there's no capture signal.